+++ src/test/java/com/braintreegateway/integrationtest/DocumentUploadIT.java
@@ -93,4 +93,19 @@ public class DocumentUploadIT extends IntegrationTest {
 
     assertEquals(ValidationErrorCode.DOCUMENT_UPLOAD_FILE_IS_TOO_LARGE, error.getCode());
   }
+
+  @Test
+  public void createReturnsErrorWhenPDFFileIsOver50Pages() {
+    URL fileToUpload = getClass().getClassLoader().getResource("fixtures/too_long.pdf");
+    DocumentUploadRequest uploadRequest = new DocumentUploadRequest(DocumentUpload.Kind.EVIDENCE_DOCUMENT,
+        new File(fileToUpload.getFile()));
+
+    Result<DocumentUpload> uploadResult = gateway.documentUpload().create(uploadRequest);
+    ValidationError error = uploadResult.getErrors()
+      .forObject("documentUpload")
+      .getAllValidationErrors()
+      .get(0);
+
+    assertEquals(ValidationErrorCode.DOCUMENT_UPLOAD_FILE_IS_TOO_LONG, error.getCode());
+  }
 }